move company resolver functions into BusinessResolver module
