+++ lib/sales_reg_web/graphql/resolvers/business_resolver.ex
@@ -1,6 +1,24 @@
 defmodule SalesRegWeb.GraphQL.Resolvers.BusinessResolver do
   use SalesRegWeb, :context
 
+  def register_company(%{user: user_id, company: company_params}, _resolution) do
+    with {:ok, company} <- Business.create_company(user_id, company_params),
+         {:ok, _} <- Business.send_registration_email(user_id, company) do
+      {:ok, company}
+    else
+      {:error, changeset} -> {:error, changeset}
+    end
+  end
+
+  def update_company(%{id: company_id, company: company_params}, _res) do
+    {_status, result} = Business.update_company_details(company_id, company_params)
+
+    case result do
+      %Company{} -> {:ok, result}
+      %Ecto.Changeset{} -> {:error, result}
+    end
+  end
+
   def upsert_bank(%{bank: params, bank_id: id}, _res) do
     Business.get_bank(id)
     |> Business.update_bank_details(params)
